Refactor utils.strings to be more logically organised
