+++ kaztron/cog/quotedb/quotecog.py
@@ -12,10 +12,11 @@ from kaztron.driver.pagination import Pagination
 from kaztron.theme import solarized
 from kaztron.driver import database as db
 from kaztron.utils.checks import mod_only, mod_channels
-from kaztron.utils.discord import Limits, user_mention, extract_user_id
+from kaztron.utils.discord import Limits, user_mention, extract_user_id, get_command_str, \
+    get_help_str, get_usage_str
 from kaztron.utils.logging import message_log_str
-from kaztron.utils.strings import format_list, get_help_str, get_timestamp_str, \
-    parse_keyword_args, get_command_str, get_usage_str, format_datetime
+from kaztron.utils.strings import format_list, parse_keyword_args
+from kaztron.utils.datetime import format_datetime, format_timestamp
 
 from kaztron.cog.quotedb.model import Quote
 from kaztron.cog.quotedb import controller as c, model
