+++ kaztron/cog/spotlight.py
@@ -17,13 +17,12 @@ from kaztron.config import get_kaztron_config, get_runtime_config
 from kaztron.driver import gsheets
 from kaztron.utils.checks import mod_only
 from kaztron.utils.converter import NaturalDateConverter
-from kaztron.utils.datetime import utctimestamp
+from kaztron.utils.datetime import utctimestamp, format_date
 from kaztron.utils.decorators import error_handler
 from kaztron.utils.discord import get_named_role, MSG_MAX_LEN, Limits, remove_role_from_all, \
-    extract_user_id, user_mention, get_member
+    extract_user_id, user_mention, get_member, get_help_str
 from kaztron.utils.logging import message_log_str, tb_log_str, exc_log_str
-from kaztron.utils.strings import format_list, get_help_str, natural_truncate, split_chunks_on, \
-    format_date
+from kaztron.utils.strings import format_list, natural_truncate, split_chunks_on
 
 logger = logging.getLogger(__name__)
