+++ kaztron/cog/sprint.py
@@ -17,11 +17,11 @@ from kaztron.config import get_kaztron_config, get_runtime_config
 from kaztron.errors import UnauthorizedUserError
 from kaztron.theme import solarized
 from kaztron.utils.checks import in_channels_cfg
-from kaztron.utils.datetime import utctimestamp
+from kaztron.utils.datetime import utctimestamp, format_date, format_timedelta
 from kaztron.utils.decorators import task_handled_errors
-from kaztron.utils.discord import check_mod, get_named_role, remove_role_from_all
+from kaztron.utils.discord import check_mod, get_named_role, remove_role_from_all, get_help_str
 from kaztron.utils.logging import message_log_str
-from kaztron.utils.strings import get_help_str, format_list, format_timedelta, format_date
+from kaztron.utils.strings import format_list
 
 logger = logging.getLogger(__name__)
