+++ kaztron/cog/wordfilter.py
@@ -9,10 +9,11 @@ from kaztron.config import get_kaztron_config, get_runtime_config
 from kaztron.driver.wordfilter import WordFilter as WordFilterEngine
 from kaztron.kazcog import ready_only
 from kaztron.utils.checks import mod_only, mod_channels
-from kaztron.utils.discord import check_role, MSG_MAX_LEN, Limits
+from kaztron.utils.discord import check_role, MSG_MAX_LEN, Limits, get_command_str, get_help_str
 from kaztron.utils.logging import message_log_str
-from kaztron.utils.strings import format_list, get_command_str, get_help_str, get_timestamp_str, \
-    split_code_chunks_on, natural_truncate
+from kaztron.utils.strings import format_list, split_code_chunks_on, natural_truncate
+
+from kaztron.utils.datetime import format_timestamp
 
 logger = logging.getLogger(__name__)
 
@@ -128,7 +129,7 @@ class WordFilter(KazCog):
                 em.set_author(name=self.match_headings[filter_type])
                 em.add_field(name="User", value=message.author.mention, inline=True)
                 em.add_field(name="Channel", value=message.channel.mention, inline=True)
-                em.add_field(name="Timestamp", value=get_timestamp_str(message), inline=True)
+                em.add_field(name="Timestamp", value=format_timestamp(message), inline=True)
                 em.add_field(name="Match Text", value=match_text, inline=True)
                 em.add_field(name="Content",
                              value=natural_truncate(message_string, Limits.EMBED_FIELD_VALUE),
