+++ kaztron/utils/logging.py
@@ -3,7 +3,7 @@ import traceback
 import logging
 
 import discord
-from kaztron.utils.strings import get_timestamp_str
+from kaztron.utils.datetime import format_timestamp
 
 
 def setup_logging(logger, config):
@@ -44,7 +44,7 @@ def message_log_str(message: discord.Message) -> str:
     [2012-01-01 12:34:56] <#channel:username#1234> "Message here"
     """
     return "[{}] <#{!s}:{!s}> {!r}"\
-        .format(get_timestamp_str(message),
+        .format(format_timestamp(message),
                 message.channel, message.author, message.content)
